Use 'snapshot' suffix for version
In Maven, you only use non-snapshot named versions for your tagged release,
'-snapshot' indicates that this is just the latest snapshot in a series of
code development. Without the '-snapshot' suffix, the artifact is assumed
to be immutable, and unique for that version.

The '1.3.0-SNAPSHOT' versions are what comes before you releasing the final
'1.3.0' release. I'm nudging the version back here so as not to confuse
the maven release plugin.